@@ -1,7 +1,7 @@
 {
   "name": "clark",
   "displayName": "CLARK: Cybersecurity Labs and Resource Knowledge-base",
-  "version": "4.29.33",
+  "version": "5.0.27",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -12,8 +12,13 @@
     "test:lighthouse:debug": "node --inspect node_modules/.bin/jest --watch -c ./lighthouse/jest.config.lighthouse.js --runInBand",
     "lint": "npx ng lint clark",
     "e2e": "ng e2e",
-    "storybook": "start-storybook"
+    "storybook": "start-storybook",
+    "prepare": "husky install"
   },
+  "@comment dependencies": [
+    "DO NOT UPDATE ng2-ckeditor",
+    "typescript should be gradually stepped up -- stable as of 2022_09_16"
+  ],
   "dependencies": {
     "@angular/animations": "^13.1.2",
     "@angular/cdk": "^13.1.2",
@@ -33,9 +38,9 @@
     "aws-sdk": "^2.1058.0",
     "core-js": "^3.6.5",
     "md5": "^2.2.1",
-    "ng-click-outside": "^4.0.0",
+    "ng-click-outside": "^9.0.1",
     "ng2-charts": "^2.3.0",
-    "ng2-ckeditor": "^1.3.6",
+    "ng2-ckeditor": "1.3.6",
     "ngx-cookie": "^5.0.2",
     "ngx-cookie-service": "^13.2.0",
     "ngx-virtual-scroller": "^4.0.3",
@@ -46,11 +51,11 @@
     "tslib": "^2.3.1",
     "util": "^0.12.4",
     "uuid": "7.0.3",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.11.8"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^13.1.4",
-    "@angular-eslint/builder": "12.7.0",
+    "@angular-eslint/builder": "^13.5.0",
     "@angular-eslint/eslint-plugin": "12.7.0",
     "@angular-eslint/eslint-plugin-template": "12.7.0",
     "@angular-eslint/schematics": "^13.0.1",
@@ -72,7 +77,7 @@
     "eslint-plugin-import": "latest",
     "eslint-plugin-jsdoc": "^37.6.1",
     "eslint-plugin-prefer-arrow": "latest",
-    "husky": "^7.0.0",
+    "husky": "^8.0.1",
     "jasmine-spec-reporter": "~5.0.0",
     "jest": "^27.4.7",
     "jest-preset-angular": "^11.0.1",
@@ -81,7 +86,7 @@
     "sass": "^1.46.0",
     "ts-jest": "^27.1.3",
     "tslib": "^1.10.0",
-    "typescript": "4.5.4"
+    "typescript": "4.6.2"
   },
   "prettier": {
     "singleQuote": true
