@@ -4,33 +4,31 @@ import { Routes, RouterModule } from '@angular/router';
 import { LoginComponent } from './login/login.component';
 import { RegisterComponent } from './register/register.component';
 import { ForgotPasswordComponent } from './forgot-password/forgot-password.component';
-import { ResetPasswordComponent } from './reset-password/reset-password.component';
-import { CanResetPasswordGuard } from './can-reset-password.guard';
+import { ChangePasswordComponent } from './change-password/change-password.component';
 import { AuthComponent } from './auth.component';
 import { EmailVerifiedComponent } from './email-verified/email-verified.component';
-import { CookieGuard } from 'app/core/cookie.guard';
-// eslint-disable-next-line @typescript-eslint/naming-convention
-const auth_routes: Routes = [
+import { CanChangePasswordGuard } from './can-change-password.guard';
+
+const AuthRoutes: Routes = [
   {
     path: '',
     component: AuthComponent,
     children: [
-      { path: '', redirectTo: 'login', pathMatch: 'full' },
       { path: 'login', component: LoginComponent, data: { title: 'Login' } },
       {
         path: 'register',
         component: RegisterComponent,
         data: { title: 'Register' },
-        canActivate: [CookieGuard],
       },
       { path: 'forgot-password', component: ForgotPasswordComponent, data: { title: 'Change Password'} },
       {
-        path: 'reset-password',
-        component: ResetPasswordComponent,
-        canActivate: [CanResetPasswordGuard],
-        data: { title: 'Reset-Password'}
+        path: 'change-password',
+        component: ChangePasswordComponent,
+        data: { title: 'Reset-Password'},
+        canActivate: [CanChangePasswordGuard]
       },
       { path: 'email-verified', component: EmailVerifiedComponent, data: { title: 'Email Verfication'} },
+      { path: '', redirectTo: 'login', pathMatch: 'full' },
       // Catch All
       { path: '**', redirectTo: 'login', pathMatch: 'full' }
     ]
@@ -38,7 +36,7 @@ const auth_routes: Routes = [
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(auth_routes)],
+  imports: [RouterModule.forChild(AuthRoutes)],
   exports: [RouterModule]
 })
 export class AuthRoutingModule {}
