@@ -1,38 +1,19 @@
-import { waitForAsync, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EmailVerifiedComponent } from './email-verified.component';
-import { AuthService } from 'app/core/auth.service';
-import { NavbarService } from 'app/core/navbar.service';
 
-const promise = {
-  then: (callback) => callback(),
-  catch: (error) => promise,
-};
-
-const authServiceStub = {
-  validateAndRefreshToken: (callback) => promise,
-  refreshToken: () => promise,
-};
-const navbarServiceStub = {
-  hide: () => {},
-};
-
-describe('ForgotPasswordComponent', () => {
+describe('EmailVerifiedComponent', () => {
   let component: EmailVerifiedComponent;
   let fixture: ComponentFixture<EmailVerifiedComponent>;
 
   beforeEach(async () => {
-    TestBed.configureTestingModule({
-    declarations: [EmailVerifiedComponent],
-    providers: [
-        { provide: AuthService, useValue: authServiceStub },
-        { provide: NavbarService, useValue: navbarServiceStub },
-    ],
-    teardown: { destroyAfterEach: false }
-}).compileComponents();
+    await TestBed.configureTestingModule({
+      declarations: [ EmailVerifiedComponent ]
+    })
+    .compileComponents();
   });
 
-  beforeEach(async () => {
+  beforeEach(() => {
     fixture = TestBed.createComponent(EmailVerifiedComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
@@ -41,16 +22,4 @@ describe('ForgotPasswordComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
-
-  it('should set loading to false when component initialization completes', fakeAsync(() => {
-    component.ngOnInit();
-    tick();
-    expect(component.isLoading).toBeFalsy();
-  }));
-
-  it('should set hasValidToken to true when component initialization completes', fakeAsync(() => {
-    component.ngOnInit();
-    tick();
-    expect(component.hasValidToken).toBeTruthy();
-  }));
 });
