@@ -13,7 +13,7 @@ import { environment } from '@env/environment';
 import { ToastrOvenService } from './shared/modules/toaster/notification.service';
 import { CookieAgreementService } from './core/cookie-agreement.service';
 import { SubscriptionAgreementService } from './core/subscription-agreement.service';
-import { NavbarService } from './core/navbar.service';
+import { CookieService } from 'ngx-cookie-service';
 
 @Component({
   selector: 'clark-root',
@@ -77,12 +77,14 @@ export class ClarkComponent implements OnInit {
     private router: Router,
     private activatedRoute: ActivatedRoute,
     private titleService: Title,
+    private route: ActivatedRoute,
     private _: HistoryService,
     private messages: MessagesService,
     private toaster: ToastrOvenService,
     private view: ViewContainerRef,
     private cookieAgreement: CookieAgreementService,
-    private subscriptionAgreement: SubscriptionAgreementService
+    private subscriptionAgreement: SubscriptionAgreementService,
+    private cookies: CookieService,
   ) {
     this.isUnderMaintenance = false;
 
@@ -107,6 +109,11 @@ export class ClarkComponent implements OnInit {
 
     this.toaster.setPosition({ x: 'left', y: 'bottom' });
     this.toaster.init(this.view);
+    this.route.queryParams.subscribe(() => {
+      if (this.route.snapshot.queryParams.err) {
+        this.toaster.error( 'SSO Error', decodeURIComponent(this.route.snapshot.queryParams.err));
+      }
+    });
   }
 
   ngOnInit(): void {
@@ -132,13 +139,21 @@ export class ClarkComponent implements OnInit {
     }
 
     this.setPageTitle();
+
+    if (this.cookies.check('ssoToken')) {
+      this.authService.setSsoSession(this.cookies.get('ssoToken'));
+      const redirect = localStorage.getItem('ssoRedirect');
+      this.router.navigateByUrl(redirect);
+      localStorage.removeItem('ssoRedirect');
+    } else if(localStorage.getItem('ssoRedirect')) {
+      localStorage.removeItem('ssoRedirect');
+    }
+
   }
 
   reloadPage() {
     location.reload();
   }
-
-
 
   /**
    * Function passes cookie agreement service val to create new agreement
@@ -194,7 +209,7 @@ export class ClarkComponent implements OnInit {
             data = activeRoute.snapshot.data.title;
           }
           if (data !== undefined) {
-            this.titleService.setTitle(data + ' | CLARK');
+            this.titleService.setTitle('CLARK | ' + data);
           }
         });
       });
