@@ -13,14 +13,15 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { UnsupportedComponent } from './unsupported.component';
 import { NotFoundComponent } from './not-found.component';
 import { ScrollingModule } from '@angular/cdk/scrolling';
-import { NavbarComponent } from './components/navbar/navbar.component';
 import { CookiesComponent } from './components/cookies/cookies.component';
-import { MessageComponent } from './components/navbar/components/message/message.component';
-import { SearchComponent } from './components/navbar/components/search/search.component';
+import { MessageComponent } from './components/message/message.component';
+import { SearchComponent } from './components/search/search.component';
 import { MaintenancePageComponent } from './maintenance-page/maintenance-page.component';
 import { UnauthorizedComponent } from './unauthorized.component';
 import { FormsModule } from '@angular/forms';
 import { SubscriptionComponent } from './components/subscription/subscription.component';
+import { PrimaryNavbarComponent } from './components/primary-navbar/primary-navbar.component';
+import { SecondaryNavbarComponent } from './components/secondary-navbar/secondary-navbar.component';
 @NgModule({
   imports: [
     BrowserModule,
@@ -35,13 +36,14 @@ import { SubscriptionComponent } from './components/subscription/subscription.co
     ClarkComponent,
     UnsupportedComponent,
     NotFoundComponent,
-    NavbarComponent,
     CookiesComponent,
     MessageComponent,
     SearchComponent,
     MaintenancePageComponent,
     UnauthorizedComponent,
-    SubscriptionComponent
+    SubscriptionComponent,
+    PrimaryNavbarComponent,
+    SecondaryNavbarComponent
   ],
   bootstrap: [ClarkComponent],
   providers: [TitleCasePipe, Title, { provide: UrlSerializer, useClass: CustomUrlSerializer }]
