@@ -113,7 +113,8 @@ export class RelevancyService {
         .toPromise()
         .then(
           (res: any) => {
-            resolve(res);
+            const sorted = res.sort(( a, b ) => ( a.name > b.name ) ? 1 : (( b.name > a.name ) ? -1 : 0) );
+            resolve(sorted);
           },
           (err) => {
             reject(err);
