@@ -14,16 +14,12 @@ import { CoreModule } from 'app/core/core.module';
 // Modules
 import { CubeCoreModule } from './core/cube-core.module';
 import { CubeSharedModule } from './shared/cube-shared.module';
-import { HomeModule } from './home/home.module';
 import { SharedModule } from '../shared/shared.module';
 
 // Components
 import { CubeComponent } from './cube.component';
 import { BrowseComponent } from './browse/browse.component';
 import { UserProfileComponent } from './user-profile/user-profile.component';
-import { UserPreferencesComponent } from './user-profile/user-preferences/user-preferences.component';
-import { UserInformationComponent } from './user-profile/user-information/user-information.component';
-import { UserEditInformationComponent } from './user-profile/user-edit-information/user-edit-information.component';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatListModule } from '@angular/material/list';
@@ -54,6 +50,11 @@ import { LibraryModule } from './library/library.module';
 import { AboutClarkComponent } from './content-pages/about-us/about-us.component';
 import { ContributePageComponent } from './content-pages/contribute-page/contribute-page.component';
 import { EditorialProcessComponent } from './content-pages/editorial-process/editorial-process.component';
+import { ProfileLearningObjectsComponent } from './user-profile/components/profile-learning-objects/profile-learning-objects.component';
+import { ProfileHeaderComponent } from './user-profile/components/profile-header/profile-header.component';
+import { EditProfileComponent } from './user-profile/components/edit-profile/edit-profile.component';
+import { HomeModule } from './home/home.module';
+import { AboutPhilosophyComponent } from './content-pages/about-us/philosophy/philosophy.component';
 
 
 /**
@@ -66,9 +67,6 @@ import { EditorialProcessComponent } from './content-pages/editorial-process/edi
     CubeComponent,
     BrowseComponent,
     UserProfileComponent,
-    UserPreferencesComponent,
-    UserInformationComponent,
-    UserEditInformationComponent,
     CollectionsComponent,
     TermsOfServiceComponent,
     FilterComponent,
@@ -86,7 +84,11 @@ import { EditorialProcessComponent } from './content-pages/editorial-process/edi
     GuidelineFilterComponent,
     AboutClarkComponent,
     ContributePageComponent,
-    EditorialProcessComponent
+    EditorialProcessComponent,
+    ProfileLearningObjectsComponent,
+    ProfileHeaderComponent,
+    EditProfileComponent,
+    AboutPhilosophyComponent,
   ],
   schemas: [
     NO_ERRORS_SCHEMA,
@@ -106,7 +108,6 @@ import { EditorialProcessComponent } from './content-pages/editorial-process/edi
     CubeSharedModule,
     CubeRoutingModule,
     SharedModule,
-    HomeModule,
     ModalModule,
     CollectionModule,
     AboutUsModule,
@@ -115,7 +116,8 @@ import { EditorialProcessComponent } from './content-pages/editorial-process/edi
     MatTabsModule,
     MatExpansionModule,
     MatListModule,
-    MatExpansionModule
+    MatExpansionModule,
+    HomeModule
   ],
   providers: [LearningObjectService]
 })
