@@ -4,13 +4,11 @@ import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { BrowseComponent } from './browse/browse.component';
 import { UserProfileComponent } from './user-profile/user-profile.component';
-import { UserPreferencesComponent } from './user-profile/user-preferences/user-preferences.component';
 import { AuthGuard } from '../core/auth-guard.service';
 import { CubeComponent } from './cube.component';
 import { CollectionsComponent } from './collections/collections.component';
 import { CollectionDetailsComponent } from './collection-details/collection-details.component';
-import { ProfileGuard } from './core/profile.guard';
-import { UserResolver } from './core/user.resolver';
+import { ProfileResovler } from './core/profile.resolver';
 import { TermsOfServiceComponent } from './terms-of-service/terms-of-service.component';
 import { AccessibilityStatementComponent } from './accessibility-statement/accessibility-statement.component';
 import { PressComponent } from './press/press.component';
@@ -20,6 +18,8 @@ import { DonateComponent } from './donate/donate.component';
 import { AboutClarkComponent } from './content-pages/about-us/about-us.component';
 import { ContributePageComponent } from './content-pages/contribute-page/contribute-page.component';
 import { EditorialProcessComponent } from './content-pages/editorial-process/editorial-process.component';
+import { NotFoundComponent } from 'app/not-found.component';
+import { CollectionsGuard } from 'app/core/collections.guard';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 const cube_routes: Routes = [
@@ -29,7 +29,7 @@ const cube_routes: Routes = [
     children: [
       { path: 'home', component: HomeComponent, data: { title: 'Home', hideTopBar: 'true'} },
       { path: '', redirectTo: '/home', pathMatch: 'full' },
-      { path: 'c/:abvName', component: CollectionDetailsComponent },
+      { path: 'c/:abvName', component: CollectionDetailsComponent, canActivate: [CollectionsGuard] },
       { path: 'c', component: CollectionsComponent },
       {
         path: 'about',
@@ -87,13 +87,8 @@ const cube_routes: Routes = [
         path: 'users/:username',
         component: UserProfileComponent,
         resolve: {
-          user: UserResolver
+          user: ProfileResovler
         }
-      },
-      {
-        path: 'users/:username/preferences',
-        component: UserPreferencesComponent,
-        canActivate: [AuthGuard]
       },
       {
         path: 'editorial-process',
@@ -111,9 +106,8 @@ const cube_routes: Routes = [
       // Catch All
       {
         path: '**',
-        component: UserProfileComponent,
+        component: NotFoundComponent,
         pathMatch: 'full',
-        canActivate: [ProfileGuard]
       },
     ]
   },
