@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { AuthService } from './core/auth.service';
 
 @Component({
   selector: 'clark-unauthorized',
@@ -11,11 +10,12 @@ export class UnauthorizedComponent implements OnInit {
 
   statusCode: string;
   redirectUrl: string;
-  constructor(private route: ActivatedRoute, private auth: AuthService) { }
+  constructor(private route: ActivatedRoute) { }
 
   ngOnInit() {
-    this.statusCode = this.route.snapshot.paramMap.get('code');
-    this.redirectUrl = this.route.snapshot.paramMap.get('redirect');
+    this.route.paramMap.subscribe((params) => {
+      this.statusCode = params.get('code');
+      this.redirectUrl = params.get('redirect');
+    });
   }
-
 }
